Added rounding and borders to popups
